CuotaUpdate

+++ CuotaAnual/Cuota.vb
@@ -1,7 +1,7 @@
 ﻿Public Class Cuota
-    Public Property Capital As Integer = 1800000.0
+    Public Property Capital As Integer = 1800000
     Public Property Plazo As Integer = 30
-    Public Property Interes As Integer = 0.012
+    Public Property Interes As Single = 0.012
 
     Public Function cuotaAnual() As Single
         cuotaAnual = Capital * Interes / (1 - (1 + Interes) ^ -Plazo)
